requirements.txt
Flask==2.1.3
requests
flask_cors
notion-client
urllib3
pyjwt
flask[async]
gunicorn
python-dotenv